requirements.txt
qiskit
matplotlib
azure-quantum[qiskit]